wip:refactor to typescript
